requirements.txt
mistralai>=1.0.0
kaggle>=1.6.0
huggingface_hub>=0.20.0
ddgs>=9.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
rich>=13.0.0
python-dotenv>=1.0.0
scholarly>=1.7.0
